refactored

+++ src/main/java/com/github/cpo1964/platform/selenium/SeleniumInterface.java
@@ -264,7 +264,7 @@ public interface SeleniumInterface extends ReportInterface {
 	 * @param timeout the timeout
 	 * @return true, if successful
 	 */
-	boolean exists(String locatorDelegate, int timeout);
+	boolean exists(String locatorDelegate, long timeout);
 
 	/**
 	 * Exists.
